Split logic for map controls ui
